miniserver.c: Fix compiler warning

+++ upnp/src/genlib/miniserver/miniserver.c
@@ -1466,7 +1466,7 @@ int StopMiniServer()
 		ssdpAddr.sin_port = htons(miniStopSockPort);
 		sendto(sock,
 			buf,
-			(int)bufLen,
+			bufLen,
 			0,
 			(struct sockaddr *)&ssdpAddr,
 			socklen);
